Added server tests via supertest
