+++ server/app.js
@@ -1,0 +1,18 @@
+const express = require('express');
+const bodyparser = require('body-parser');
+const db = require('../database/queries.js');
+
+const app = express();
+app.use(bodyparser());
+app.use(bodyparser.json());
+app.use(bodyparser.urlencoded({ extended: true }));
+
+app.use(express.static('public/'));
+
+app.get('/listing_info', db.getInfo);
+
+app.get('/', (req, res) => {
+  res.status(200).send('Hello World');
+});
+
+module.exports = app;
